dynamically load the ado-file
