+++ packages/ado-extension/src/ado-extension.ts
@@ -1,10 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-try {
-    const url = process.argv.slice(2)[0];
-    console.log(`Scanning ${url}`);
-} catch (error) {
-    console.log('Exception thrown in action: ', error);
-    process.exit(1);
+import * as adoTask from 'azure-pipelines-task-lib/task';
+
+export function runScan() {
+    try {
+        const url = adoTask.getInput('url', true);
+        console.log(`Scanning ${url}`);
+    } catch (error) {
+        console.log('Exception thrown in action: ', error);
+        process.exit(1);
+    }
 }
